+++ src/cmds.h
@@ -1,0 +1,6 @@
+#ifndef __KAD_CMDS_H__
+#define __KAD_CMDS_H__
+
+extern struct cmd_def* cmd_defs[];
+
+#endif
